backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# Audio Processing
openai-whisper
soundfile
torchaudio

# Deep Learning
torch
transformers
numpy

# Text-to-Speech
gTTS

# LLM
google-generativeai

# Configuration
python-dotenv

# Type hints
pydantic
